app/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
filelock==3.12.3
idna==3.4
platformdirs==3.10.0
python-dotenv
requests==2.31.0
urllib3==2.0.4
virtualenv==20.24.4